# swagger/swagger.yml
swagger: "2.0"
info:
  title: Users Info
  description: Information about users
  version: 1.0.0
host: localhost
schemes:
  - http
basePath: /v1
consumes:
  - application/json
produces:
  - application/json

paths:
  /users/:
    get:
      summary: Get all users
      operationId: GetUsers
      parameters:
        - $ref: '#/parameters/user_id'
        - $ref: '#/parameters/first_name'
        - $ref: '#/parameters/second_name'
        - $ref: '#/parameters/email'
        - $ref: '#/parameters/country_code'
      responses:
        200:
          description: General information about users
          schema:
            $ref: '#/definitions/ArrayOfUsers'
        400:
          description: Bad request
          schema:
            $ref: '#/definitions/Error'
        500:
          description: Internal server error
          schema:
            $ref: '#/definitions/Error'

  /user/{user_id}:
    get:
      summary: Get user by id
      operationId: GetUserById
      parameters:
        - $ref: '#/parameters/user_id_in_path'
      responses:
        200:
          description: General information about user
          schema:
            $ref: '#/definitions/User'
        400:
          description: Bad request
          schema:
            $ref: '#/definitions/Error'
        500:
          description: Internal server error
          schema:
            $ref: '#/definitions/Error'


parameters:
  user_id:
    name: user_id
    in: query
    description: User identificator
    type: string

  user_id_in_path:
    name: user_id
    in: path
    description: User identificator
    type: string
    required: true

  first_name:
    name: first_name
    in: query
    description: User first_name
    type: string

  second_name:
    name: second_name
    in: query
    description: User second_name
    type: string


  country_code:
    name: country_code
    in: query
    description: User country_code
    type: string

  email:
    name: email
    in: query
    description: User email
    type: string

definitions:
  Error:
    type: object
    required:
      - message
      - code
    description: Error message
    properties:
      message:
        type: string
      code:
        type: integer

  ArrayOfUsers:
    type: array
    items:
      $ref: '#/definitions/User'

  User:
    type: object
    required:
      - id
      - firstname
      - secondname
      - email
      - countrycode
    properties:
      id:
        type: string
        description: User's identificator
      firstname:
        type: string
        description: First name
      secondname:
        type: string
        description: Second name
      email:
        type: string
        description: Email
      countrycode:
        type: string
        description: Country code